Separate the directly defined and derived variables
There is an open issue https://github.com/jgthms/bulma/issues/187
which explains most of this problem. Basically, one wants to override
some values, set some new values, but also use some other defined values
that are part of the framework as new derived variables.

Specfically, the example in the issue has

```
@import "bulma/sass/utilities/variables"

$strawberry: #BF3030

$primary: $grey-dark
$warning: $strawberry
$info: $grey-light

@import "bulma"
```

which is now possible to do with this PR. The docs and the comments in the PR to do something like

```
$primary: $grey-dark
$warning: $strawberry

$info: $grey-light
$success: $green
$danger: $orange

@import "bulma/bulma"
```

does not work because of undefined variables.