Active Bulma - Jump start your Bulma creation.
ActiveBulma is a simple yet powerful Javascript implementation for most of bulma's components that needs javascript in order to interact with them. just drop it in your project and Your good to go with modal, dropdown, and the navbar burger. 
In order to to use the modal, call the "openModal()" function. everything ese just works out of the box.