Update generic.sass
updated $body-overflow-y to 'auto'. 
The previous value  for overflow-y was 'scroll' which causes a scroll bar to appear even in there is no need to scroll.  When the property is set to auto, the scroll bar will automatically appear if there is an overflow.
Having a scrollbar when it is not need is annoying and not a good look to a website.